feat: rename project to lowercase, restructure main page with routing, and add SummaryList component
