+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "DictatorWatch.info",
+  "name": "dictatorwatch-info",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
